requirements.txt
flask
mysql-connector-python
pytest
gunicorn==20.1.0